requirements.txt
Django==6.0.1
djangorestframework==3.16.1
drf-spectacular==0.29.0
djangorestframework_simplejwt==5.5.1
black==26.1.0
isort==7.0.0
flake8==7.3.0